INC-936: manually set whenUpdated in NextReviewDate records
`@LastModifiedDate` annotation doesn't work. It's probably because we're
not using spring-data-jpa and instead we use `spring-data-r2dbc`.

Only place which updates `NextReviewDate` records is the `NextReviewDateUpdaterService`
so we can explicitly set the `whenUpdated` field to now when this happens.

Tests had to be updated in order for some mocks to work as expected but
overall this is simple enough and it works.