Attempt to fix flaky integration test
The test seems to fail quite frequently, example:

https://app.circleci.com/pipelines/github/ministryofjustice/hmpps-incentives-api/1152/workflows/a3fb2a68-8fe0-417d-bdb6-10a933ab2c88/jobs/3838

The tests tend to fail when waiting for the request to Prison API in an
attempt to determine whether the admission domain event(s) have been
processed.

Maybe these tests starting to early causing the 30s timeout to lapse.

I've added another condition to check a review record has been created
before those prisonApi mock requests are checked: An admission creates
a review record so check for new records should be a good indication
the event may have been processed (or at least almost).

I'm not 100% sure this will fix this test as it's still odd that 30s
is not enough for this event to be processed.