INC-927: Publish domain event when next review date changes (#295)
* INC-927: Publish event when next review date changes

Publish a `incentives.prisoner.next-review-date-changed` domain event
when a next review date is updated and it has changed.

An event per prisoner is published. The event is only published when
we're sure the next review date changed (e.g. when the record changed
rather than when the record was created)

NOTES:
1. ideally we should publish these multple domain events concurrently
   rather sequentially. Waiting for many domain events to be published
   could be a problem
2. we could include the new next review date in the domain event message
   (but this requires some changes to the `SnsService#publishDomainEvent()`)

* INC-927: Publish 'next-review-date-changed' events concurrently

When a bunch of next review dates has changed we publish these domain
events concurrently instead of publishing them one by one (sequencially)

* INC-927: Include `nextReviewDate` in the domain event

When a next review date changes we publish a 'incentives.prisoner.next-review-date-changed'
domain event. This now includes the new `nextReviewDate` in the `additionalInformation`.

NOTE: I had to refactor the `publishDomainEvent()` method (in `SnsService`)
to be more flexible and accept the `additionalInformation` object as an
argument instead of instanciating it inside the method. This means
all the existing method calls had to change from something like

```kotlin
snsService.publishDomainEvent(id, prisonerNumber, occurredAt, eventType, description)
```

to

```kotlin
snsService.publishDomainEvent(eventType, description, occurredAt, additionalInformation)
```

(I had to change the parameters anyway so I've rearranged the order to have
`additionalInformation` last)

* INC-927: Updated 'next-review-date-changed' event description

As per PR comment.