INC-957: Next review date services don't use Offender Search API (directly)
When getting/updating the next review date for a single prisoner the
services had to get the extra information needed for the calculation.

This was previously done by going to Offender Search API.
However when re-indexing a new prisoner some of the data may be missing in
Offender Search API/ElasticSearch, causing all sort of exceptions.

The services are now getting the required data from Prison API (at the
source) which should avoid the cycle issues.

NOTE: Syscon said the `GET /api/bookings/:bookingId?extraInfo=true` should
be relatively quick and it shouldn't be causing slowness. Furthermore
we're now making only 1 request to Prison API instead of 2 requests,
(one to Prison API to get the basic info/offender number, another to
Offender Search API) which may actually make this faster on top of being
simpler.