INC-821: Added READMISSION review type
This is to represent reviews triggered when a prisoner is received
(`prisoner-offender-search.prisoner.received` event) with reason
`READMISSION`. This happens when a prisoner is readmissed with their
existing bookingId.

NOTE: This could happen for a number of reasons and we cannot 100%
guarantee a readmission is a recall because someone could be readmissed
to prison for a number of other reasons (e.g. remand).

This is also why I didn't name the review type "RECALL", because it would
be potentially misleading.