INC-961: Fix default value for sync mechanism after rename
The environment variable was renamed in most places but not in the default
values leading to the pod having both environment variables set but
with different values. I don't know which one would win and it's best to
avoid confusion anyway.

See Pod variables in `dev`:

```YAML
    - name: FEATURE_REVIEWADDEDSYNCMECHANISM
      value: PRISON_API
    - name: FEATURE_REVIEW_ADDED_SYNC_MECHANISM
      value: DOMAIN_EVENT
```

(which one will Spring Boot use?)