INC-821: Next review date rule for readmissions
If the last "review" is of type "READMISSION" we apply the same rule
we use for new prisoners but with a starting date of when the prisoner
was readmitted.

NOTE: This covers the recall rule in incentives policy but because we
cannot discriminate recalls readmissions from other types of readmissions
(e.g. remands) it's considering any prisoner being readmitted as a new
prisoner.