INC-958: Process 'alerts-updated' domain events to update next review date
These (`prisoner-offender-search.prisoner.alerts-updated`) were added later
and they're more specific than the older `prisoner-offender-search.prisoner.alerts-updated`.

The structure of the domain events is different but the good news is that
we can check in the domain event if an ACCT alert has been added or removed.
And we also get the bookingId, so we can avoid

NOTE: I've removed the old message so this can only be merged once the
CP PRs are approved/merged