INC-821: Next review date calculation takes readmissions in consideration (#303)
* INC-821: Increased `review_type` column length

Limit was 10 characters but 'READMISSION' is 11 characters and it doesn't
fit. Increased to 16 characters.

* INC-821: Added READMISSION review type

This is to represent reviews triggered when a prisoner is received
(`prisoner-offender-search.prisoner.received` event) with reason
`READMISSION`. This happens when a prisoner is readmissed with their
existing bookingId.

NOTE: This could happen for a number of reasons and we cannot 100%
guarantee a readmission is a recall because someone could be readmissed
to prison for a number of other reasons (e.g. remand).

This is also why I didn't name the review type "RECALL", because it would
be potentially misleading.

* INC-821: Different review type for readmissions

Before this admissions and readmissions are treated the same and the
resulting "review" has a review type of `ADMISSION`.

However in order to calculate next review dates for recalls (a special
case of readmissions with same booking) we need to discriminate between
first time admissions and readmissions.

* INC-821: Next review date rule for readmissions

If the last "review" is of type "READMISSION" we apply the same rule
we use for new prisoners but with a starting date of when the prisoner
was readmitted.

NOTE: This covers the recall rule in incentives policy but because we
cannot discriminate recalls readmissions from other types of readmissions
(e.g. remands) it's considering any prisoner being readmitted as a new
prisoner.