INC-988: Cache the global list of all incentive levels
This practically almost never change and we use it in quite a few places
so it's wasteful to make an HTTP request every time we need this list.

I tried to use Spring Boot's caching mechanism but for some reason I
couldn't get it to work - maybe is the combination of libraries we use?

The value is cached for 24 hours in memory, it's very simple but it should
be good enough for this use case.