INC-1055: Updated 'Reviews' endpoint format with counts
As discussed with Igor (see ticket) we can return the list of levels for
the prison with the counts for each level. This is nicer and it also
meand the UI will be able to stop making a request to get this list.

Example:

```JSON
{
  "levels": [
    {
      "levelCode": "BAS",
      "levelName": "Basic",
      "reviewCount": 1,
      "overdueCount": 0
    },
    {
      "levelCode": "STD",
      "levelName": "Standard",
      "reviewCount": 3,
      "overdueCount": 0
    },
    {
      "levelCode": "ENH",
      "levelName": "Enhanced",
      "reviewCount": 1,
      "overdueCount": 1
    }
  ],
  // [...]
}
```

**NOTE**: The existing `reviewCount` field which is the number of all unpaginated
reviews for the given level is now deprecated because this information
is now in the `levels` field.

**NOTE**: The `overdueCount` field which is the number of all ovedue
reviews (regardless of the level) is now deprecated because this information
can be derived by summing all the `overdueCount` values in the `levels`
list. Furthermore the latest design doesn't show the total number of
overdue above the reviews table anymore so this shouldn't be necessary
anyway.