INC-1164: Removed /incentives-summary resource
This only contained `GET /incentives-summary/prison/{prisonId}/location/{locationId}`
endpoint used only by the Incentives UI.

On Monday 3rd April we'll roll out the improved 'Reviews' table to all
prisons. This uses a different endpoint so we can remove the old one
and the related code to simplify things.
This will be particularly useful when we'll plug in the new reference data
instead of using Prison API as we currently do (as it will mean less code
will need to be updated)

**NOTE**: This commit only removes the resource class and its tests, other
classes and functions will need to go as well as they were only used by
this resource