INC-1164: Made `IepDetail`'s `id` non-nullable
This is a representation of an Incentive level record for a prisoner.
The `id` was nullable because this was coming from NOMIS/Prison API
during the transition period. This was only present when data was coming
from our database.

However this data always come from our database now so the `id` will be
always present (this happen to be the `prisoner_iep_level`'s primary key)

By making this mandatory we may be able to simplify some code but also
reduce confusion.

NOTE: This is only used for some of the responses format and never as
format for requests body, so it shouldn't break any compatibility.