INC-1164: Made `IepDetail`'s `reviewType` non-nullable
This is a representation of an Incentive level record for a prisoner.
The `reviewType` was nullable because this was coming from NOMIS/Prison API
during the transition period. This was only present when data was coming
from our database (because `reviewType` is a new concept).

However this data always come from our database now so the `reviewType` will be
always present (this is currently NULL-able in the DB but it shouldn't)

By making this mandatory we may be able to simplify some code but also
reduce confusion.

NOTE: This is only used for some of the responses format and never as
format for requests body, so it shouldn't break any compatibility.