INC-1164: Made `IepDetail`'s `prisonerNumber` non-nullable
This is a representation of an Incentive level record for a prisoner.
The `prisonerNumber` was nullable because this was coming from NOMIS/Prison API
during the transition period. This was only present when data was coming
from our database (because `prisonerNumber` is a DPS concept).

However this data always come from our database now so the `prisonerNumber` will be
always present (this column is non-NULLABLE in the DB)

By making this mandatory we may be able to simplify some code but also
reduce confusion.

NOTE: This is only used for some of the responses format and never as
format for requests body, so it shouldn't break any compatibility.