INC-1164: Made `IepDetail`'s `auditModuleName` non-nullable
`IepDetail` is a representation of an Incentive level record for a prisoner.
The `auditModuleName` was nullable because we don't store it in our
database. In practice we always set this to `INCENTIVES_API` when instanciating
an `IepDetail`, as the data would always coming from our database.

By making this mandatory we may be able to simplify some code but also
reduce confusion.

NOTE: This is only used for some of the responses format and never as
format for requests body, so it shouldn't break any compatibility.