INC-1164: Made `IepSummary`'s `id`/`prisonerNumber` non-nullable
`IepSummary` contains a summary of a prisoner current incentive level.
An `IepSummary` instance has values like `id`, `prisonerNumber`, `iepLevel`,
etc set as the same value as the current/latest Incentive review.

All places where an `IepSummary` is instanciated check is the list of
Incentive reviews is non-empty (`iepDetails` is a list of `IepDetail`).

This means `id` and `prisonerNumber` will always be available and non-nullable.

The `id`/`prisonerNumber` were nullable because historically the data
could be coming from NOMIS/Prison API. During the transition period
`id` and `prisonerNumber` could have been `null` if the data was coming
from NOMIS/Prison API.
Data is always coming from our database now so Incentive records will
always have `id` and `prisonerNumber`.

By making this mandatory we may be able to simplify some code but also
reduce confusion.

NOTE: This is only used for some of the responses format and never as
format for requests body, so it shouldn't break any compatibility.