INC-1164: Made `IepSummary`'s `nextReviewDate` non-nullable
This field was nullable because in the past the `IepSummary` data could
come from NOMIS/Prison API without details: The API callback had a
`?withDetails=` query param which when set to `false` would not return
the history of incentive records for a prisoner.

However all places where `IepSummary` are instanciated have the history
of a prisoner's incentive records so we can always calculate the `nextReviewDate`
value and we can make this mandatory.

NOTE: One of the Incentives API's endpoints still has an optional `withDetails`
query parameter which defaults to `true`. Even when `?withDetails=false`
and the historic Incentive records are not returned these are
retrieved from the DB and used to calculate `nextReviewDate`.