INC-1164: `IncentiveStoreService` removed unnecessary `bookingId` params
Both the `deleteIncentiveReview()` and the `patchIncentiveReview()`
received the `PrisonerIepLevel` to remove/update.
This already contains the `bookingId` so having that param is unnecessary
and potentially confusing.

In fact, the service would even check if the client provided an id which
doesn't match the request body and throw an exception in this case. So
it's guaranteed that the `PrisonerIepLevel` instance would contain the
correct bookingId.