`NextReviewDateService` no longer uses `IepDetail` (#402)
* NextReviewDateService no longer uses IepDetail

The class is still used, and in fact even part of the public interface
of some of the endpoints. However is kinda a legacy data class.

The `NextReviewDateService` now uses `PrisonerIepLevel` values which is
what's stored in the database. This is the new representation of an incentive
record.

* No need to convert lists of `IepDetail` to lists of `PrisonerIepLevel`

Now that the next review date calculation works with `PrisonerIepLevel`
directly we don't need to convert from `List<IepDetail>` anymore.

This is nice because that conversion required the list of Incentive levels
which required an additional API call to Prison API (at the moment at least).

* Moved code for Basic level into `IncentiveLevel` class

This is currently only used to calculate the next review date but it seems
natural to store it in the `IncentiveLevel`.