ADR: Rename Incentive level "description" to "name"
Reason: The current values for "description", e.g. "Standard" or "Enhanced"
are the **names** of the Incentive levels, NOT their description.

The new Incentive levels management UI will show levels descriptions
describing the level so it makes sense to rename the existing "description"
field.

NOTE: Prison Incentive levels have a field called `levelDescription`, this
has the same meaning and will be renamed as well to `levelName` (a new
`levelDescription` field will be added but it will have a different meaning)