Renamed Incentive levels' `description` field to `name`
The values for `description` are actually the name of an Incentive level.
The fact these were called "description" is confusing.
We're also planning to add an *actual* description field for UI purposes
so not renaming the existing fields would make this even more confusing.

NOTE: The Prison Incentive levels have a similar field called `levelDescription`.
This has to change accordingly as well but this change is not part of this
commit.