INC-1193 (part I): Renamed Incentive levels' "description" to name (#405)
* ADR: Rename Incentive level "description" to "name"

Reason: The current values for "description", e.g. "Standard" or "Enhanced"
are the **names** of the Incentive levels, NOT their description.

The new Incentive levels management UI will show levels descriptions
describing the level so it makes sense to rename the existing "description"
field.

NOTE: Prison Incentive levels have a field called `levelDescription`, this
has the same meaning and will be renamed as well to `levelName` (a new
`levelDescription` field will be added but it will have a different meaning)

* Renamed Incentive levels' `description` field to `name`

The values for `description` are actually the name of an Incentive level.
The fact these were called "description" is confusing.
We're also planning to add an *actual* description field for UI purposes
so not renaming the existing fields would make this even more confusing.

NOTE: The Prison Incentive levels have a similar field called `levelDescription`.
This has to change accordingly as well but this change is not part of this
commit.

* Renamed PrisonIncentiveLevel's `levelDescription` to `levelName`

Similarly to what done for the `IncentiveLevel` class, the `levelDescription`
fields here really mean "level name" so we're renaming for clarity.

Furthermore, we'll be adding an actual level description soon enough so
the old fields has to change.