Removed `DELETE /iep/sync/booking/{bookingId}/id/{id}`
This endpoint was used to sync deletion in NOMIS to our service.
Now that NOMIS screens for carrying Incentive reviews are off users cannot
add Incentive records in NOMIS anymore which means this syncronisation
is unnecessary/unused.

This Azure query can be used to determine if the endpoint is being used:

```Kusto
dependencies
| where target == "incentives-api.hmpps.service.justice.gov.uk"
| where name startswith_cs "DELETE /iep/sync"
```