Removed `POST /migration/booking/{bookingId}` endpoint
This endpoint was used to populate our database with all the historical
data from NOMIS when we started to move away from it.

The migration process was a one off process done once and then followed
by 2-way syncing (via domain events + sync endpoints)

As the migration has long been completed this endpoint is no longer necessary
nor used.

The following Azure' `dependencies` query can be used to check if anyone
is making requests to this endpoint:

```Kusto
dependencies
| where target == "incentives-api.hmpps.service.justice.gov.uk"
| where name startswith_cs "POST /iep/migration"
```