Removed `POST /iep/sync/booking/{bookingId}`
This endpoint was used to sync Incentive records created in NOMIS to
our service.
Now that NOMIS screens for carrying Incentive reviews are off users cannot
add Incentive records in NOMIS anymore which means this syncronisation
is unnecessary/unused.

This Azure query can be used to determine if the endpoint is being used:

```Kusto
dependencies
| where target == "incentives-api.hmpps.service.justice.gov.uk"
| where name startswith_cs "POST /iep/sync"
```