Removed unused `IepReviewReason` enum
It only contained the `USER_CREATED_NOMIS` value and it was only used to
allow Syscon sync service to ignore domain events for Incentive records
already present in NOMIS.

The reason was only passed when publishing domain events as a result of
request to `POST /iep/sync/booking/{bookingId}` therefore that param can
now be removed.

1. when an Incentive record was created in NOMIS
2. Syscon's sync service would syncronise to our service by making an HTTP
   request to `POST /iep/sync/booking/{bookingId}`
3. Incentive service would create the Incentive record in our domain
   and publish a domain event with `reason=USER_CREATED_NOMIS`
4. Syscon's sync service would ignore this domain event - because the
   Incentive record was coming from NOMIS in the first place
5. Other interested services could process this domain event

NOMIS Incentives screens are now off and syncronisation is not longer needed.