Renaming related to an update of an Incentive record
Igor suggested we keep around the logic related to the update and deletion
of an Incentive record (without exposing it for now).

This commit rename things like `SyncPatchRequest` (a concept specific
of the old `PATCH /iep/sync/...` endpoint) to `IncentiveReviewUpdate`.

In general things are renamed to be as much consistent as possible with
more recent code, e.g.:
- `iepTime` becomes `reviewTime` for consistency with the field in `PrisonerIepLevel`
- "Iep" is removed when there the code is not related to legacy Prison API classes
  ("Iep" is a term no longer used in the new Incentives policy)
- IepReview/IncentiveReview are replaced with "Incentive record" when it
  makes sense (a `PrisonerIepLevel` is not necessarily an Incentive "review",
  it could be a record generate on admission or transfer for example)
- etc