INC-1199: Better handling of Spring's ServerWebInputException
These exceptions happen automatically when the request format is not valid
and the request can't be converted into the correct input type/class, e.g.
- a required field is missing (e.g. `code` for an `IncentiveLevel`)
- a field has an incompatible type (e.g. something like `"undefined"`
  passed as a path variable for the `bookingId`)
- a value is not one of the ones in an enum (e.g. trying to sort prisoners
  by something not allowed)
- possibly others

In general these happen before the request reachs the part of the code
we have more control over. The important part is that these responds with
a `400 Bad Request` which is the correct thing to do in this instance.

Improvements:
- The `userMessage` gives some examples of why the request could be invalid
  to try to help
- The `developerMessage` now shows the cause of the exception.
  This is a bit technical but it's very useful to understand which
  part of the request caused the problem (see below)
- The logged error also shows the cause of the exception, this is much
  more useful to understand/investigate the problem than before
  (before could include something like "400 BAD_REQUEST "Failed to read HTTP message""
  which was misleading and almost sounded like a networking problem
  rather than a bad request missing a field)

NOTE: Don't know if it's a result of an upgrade of one of the dependencies
but this exception seems to be thrown also instead of `TypeMismatchException`
(the handler for this could be removed if this is the case)