INC-1199: Removed handling of `TypeMismatchException`
I believe this exception is covered by the handler for `ServerWebInputException`.
We have a test for what I think it's the reason why this handler was introduced,
and it's covered by that handler (`undefined` passed as `bookingId: Long`)

It's tricky to discriminate this specific exception handler in the logs
since the log message is the same as the handler for `ServerWebInputException`.

However if we see a bunch of 500s after deploying this we can easily revert
this commit.