Increase test coverage for the case of a prison being re-activated When a prison is closed, it's prison incentive levels should be deactivated. This tests the situation where the same prison is later reopened. NB: the first level that is activated must be a default according to business rules, but furthermore it must also have been the level that was _previously_ the default.
If a prison had these levels:
• BAS active
• STD active
• ENH active and default
and is closed, then it ends up with:
• BAS inactive
• STD inactive
• ENH inactive and default

To reactivate the whole prison the actions needed are:

1) activate ENH resulting in
• BAS inactive
• STD inactive
• ENH active and default

2) activate all other required levels (BAS & STD) resulting in
• BAS active
• STD active
• ENH active and default

only _then_ can the default level be changed if necessary.

Changing the default level for a prison when the current default is inactive is not always possible. Changing the default level necessarily requires a side effect of making the previous level no longer default. If that previous level was globally required, it will restrict updating it as inactive.