INC-1276: Add api endpoint to reset levels in a prison (#461)
* Define all associated information for prison incentive levels in level-policy mapping

* Add api endpoint to reset levels in a prison
This is useful when setting up a new prison. It can also be used to ensure that a prison has the required set of levels active and that Standard is the default for admission.
If in future there are domain events raised when a prison is created, this app could listen for them and automatically set up the prison.