INC-1296: Removed legacy endpoint `GET /iep/levels/{prisonId}`
This is an old endpoint to retrieve the incentive level for a prison.
However this had a legacy format and it was replaced by the newer
`GET `/incentive/prison-levels/{prisonId}`.

The endpoint has been deprecated for a long time.

AFAICT from the look of AppInsights dependency this is not used (or
at least not in the last week or two):

https://portal.azure.com#@747381f4-e81f-4a43-bf68-ced6a1e14edf/blade/Microsoft_OperationsManagementSuite_Workspace/Logs.ReactView/resourceId/%2Fsubscriptions%2Fa5ddf257-3b21-4ba9-a28c-ab30f751b383%2FresourceGroups%2Fnomisapi-prod-rg%2Fproviders%2Fmicrosoft.insights%2Fcomponents%2Fnomisapi-prod/source/LogsBlade.AnalyticsShareLinkToQuery/q/H4sIAAAAAAAAA1XNOw6DMBAE0J5TrKgT%252BwSUUboUUZQWIXsEC8a2vMY0HJ5PESnVaqQ3sxYR3sIbhlQbrQMSKHepR6amoZq9gc9cIPcushrmGEUJUmEDNS6Sz9uHopap%252FvV9N4PkWMmych5aI1Q%252FHx%252FSjKgdCpzoU8cURph8%252BRsZFxbbvoPD68pipy%252BScPCV1rSR5eObP%252Fg%252F3AGe4URUwgAAAA%253D%253D/timespan/2023-09-19T12%3A57%3A32.000Z%2F2023-10-10T12%3A57%3A32.286Z