INC-1321: OS API: Ability to pass a different `size` when getting prisoners
This is still `500` by default as before but it's now possible to pass
a different size to get more prisoners per request.

Offender Search API is fairly quick and it allows up to 3,000 results
per request.

This could be useful when getting all the prisoners across the estate
as that already makes a large number of HTTP requests to Offender Search
API.

By getting 3,000 prisoners per page we should be able to get all
prisoners in a given prison in 1 request (largest prison has a capacity
of ~2K prisoners)

This means we should be able to get all prisoners across the estate with
~125 HTTP requests.