INC-1321: Added back ORDER BY in SQL query to get overdue prisoners
The `DISTINC ON(prisoner_number)` acts similarly to a GROUP BY and then
filter out duplicates. This means the order of the rows before the filtering
may be significant.

In this specific case where we'll be interested in the count may not make
much of a difference but it's safer to keep it anyway.

Even tho there shouldn't be `NULL` review times, it's safer to add the
`NULLS LAST`.