INC-1321: Fix query to get overdue prisoners
The combination of `DISTING ON(prisoner_number)` and the
`WHERE` clause to only get overdue people meant that there was an edge
case.

If a prisoner number had multiple booking IDs and this person
had a recent review and an old review (overdue renewal) on a
different booking this would be still be counted as overdue.

The solution is to:
1. get all (distinct) prisoner numbers and their next review date
2. THEN filter to retain only the ones that are still overdue