INC-1343 Update dependencies and refactor code for clarity - localstack upgrade (#545)
* INC-1343 Update dependencies and refactor code for clarity

Code has been modified to update specific out-of-date dependencies and improve code clarity. Additionally, some unnecessary configuration files were removed, and comments were added to the codebase to improve readability and understandability. The codebase was further refactored to improve naming consistency across multiple files.

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>

* Update WebClient configurations

This commit modifies the WebClient configuration to include response timeout and alters the method used to construct the client. It also updates the application health timeout value from 20s to 2m and replaces "prison-api" with a system username in the authorized client manager.

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>

* Remove unused import in WebClientConfiguration

The ServerOAuth2AuthorizedClientExchangeFilterFunction import was removed from the WebClientConfiguration.kt file. It was not used in the code, thereby improving the cleanliness and readability of the codebase.

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>

* Refactor AWS SQS import and modify subscribe endpoint

The commit represents a cleanup in the codebase where individual imports of SQS model classes were replaced with a wildcard. It also includes a change to use the queue ARN instead of the queue URL as the endpoint when subscribing to the SNS topic in the IncentiveLevelResourceTestBase class. This adjustment ensures appropriate testing of the service integration with Amazon SNS and SQS.

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>

* KLint

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>

---------

Signed-off-by: mjwillis <michael.willis@digital.justice.gov.uk>