INC-1367: Use generic-service's DLQ CronJob
We currently have a "custom" CronJob which retries all the messages
in the DLQ (Dead Letter Queue).

However the "same" CronJob is provided by the `generic-service` helm
chart we use to deploy the service.
This CronJob has the advantae of being aware of the scheduled downtime
and avoiding trying to run this cronjob when the service is not running
(as that would not work).

This may (or may not) reduce the number of failures/alerts we see in the
DPS alert channels related to this cronjob failing.