Deprecate locationId field
This information is not reliable as it's not really
the location where the review took place but rather
the assigned living unit location of the prisoner
at the time the review was created in our service.

This field is optional and often null.

Critically in order to populate this field a request
to a deprecated Prison API endpoint is made.

In the interest of keeping things simple we decided
to deprecate this field and:
- always set new values to `null`, removing the dependency
  to the deprecated endpoint
- return null in our responses: this field is already
  nullable