Renamed `PrisonerAtLocation` to `PrisonerInfo`
This is the response from Prison API with the prisoner information.
It was only called `PrisonerAtLocation` because it also contained
the `assignedLivingUnitId` field.

This field was used to make another request to get the
prisoner location info (using a deprecated Prison API endpoint).

However `locationId` is no longer persisted so I've removed
this field and given this class a more fitting name.