Added `LocationService`
Prison API's locations endpoints are being deprecated in favour
of using Locations Inside Prison API:

https://locations-inside-prison-api-dev.hmpps.service.justice.gov.uk/swagger-ui/index.html

Incentives is only interested in the endpoint to get
a location by key (e.g. MDI-1). This is only used to populate
the name of the location in the Reviews table endpoint.

NOTE: There is probably more wiring required that is missing
in this commit.