INC-1371 (part 2): Get location description from Locations API (#598)
* Added `LocationService`

Prison API's locations endpoints are being deprecated in favour
of using Locations Inside Prison API:

https://locations-inside-prison-api-dev.hmpps.service.justice.gov.uk/swagger-ui/index.html

Incentives is only interested in the endpoint to get
a location by key (e.g. MDI-1). This is only used to populate
the name of the location in the Reviews table endpoint.

NOTE: There is probably more wiring required that is missing
in this commit.

* Added Locations mock server

this fixes the healthcheck tests

* Added Locations API container to docker-compose

* Get location info from Locations inside prison API

This is only used in the reviews endpoint to get the
"description" of the location (e.g. Houseblock 1)
to be displayed in the reviews table.

* Prison API: Removed now unused getLocation() method

This method in the `PrisonApiService` is no longer used
and it can be removed alongside the DTO class for its
response.