Added PrisonerSearch methods to find prisoner info
Both by prisonerNumber and bookingId.

NOTE: Finding a prisoner by bookingId is slightly more fiddly
because of the API lacking an endpoint to get a single prisoner info
by bookingId: We're using the "batch" `POST /prisoner-search/booking-ids`
endpoint instead but with a single bookingId.
Also, this means that the API would not responds 404 Not Found when
the given prisoner is not found but rather 200 OK with an empty array.

See:
- https://prisoner-search-dev.prison.service.justice.gov.uk/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config#/Popular/findByPrisonNumber
- https://prisoner-search-dev.prison.service.justice.gov.uk/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config#/Batch/findByIds