Continue using Prison API to get prisoner info **by bookingId**
Prisoner Search API is much faster and nicer but only contains
information about the latest bookingId (a prisoner can have
many booking ids, e.g. if they've been in prison multiple
times).

Although this may not be a problem in practive in theory a request
to our API could provide an older booking ID and that would not
work the same way if we start using Prisoner Search API.

For this reason this reverts some of the changes related to getting
information by `bookingId`.

NOTE: The `PrisonerIncentiveReviewService` class handles both
the case where a `prisonerNumber` or a `bookingId` are used.
In the past the information about a prisoner came from Prison API
and the response class was used as a generic input for the methods
to add incentive reviews.

This is no longer the case as data may come from Prisoner Search API
(when getting info by `prisonerNumber`) or from Prison API (when
getting info by `bookingId`). For this reason I've introduced an
`PrisonerBasicInfo` interface.