INC-1395: Use Prisoner Search API to get prisoner info (#683)
* Added PrisonerSearch methods to find prisoner info

Both by prisonerNumber and bookingId.

NOTE: Finding a prisoner by bookingId is slightly more fiddly
because of the API lacking an endpoint to get a single prisoner info
by bookingId: We're using the "batch" `POST /prisoner-search/booking-ids`
endpoint instead but with a single bookingId.
Also, this means that the API would not responds 404 Not Found when
the given prisoner is not found but rather 200 OK with an empty array.

See:
- https://prisoner-search-dev.prison.service.justice.gov.uk/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config#/Popular/findByPrisonNumber
- https://prisoner-search-dev.prison.service.justice.gov.uk/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config#/Batch/findByIds

* Use Prisoner Search API to get prisoner info in services

Instead of methods in `PrisonApiService`.

NOTE: More tests need fixing as they're still mocking the
Prison API requests.

* Removed now unused `prisonerAtLocation()` test helper

* Prisoner Search Wiremock: mock `getPrisonerInfo()` requests

* Fix ManageIncentiveReviewsResourceTest

By mocking Prisoner Search API requests to get prisoner info,
instead of Prison API's ones.

* Fixed PrisonOffenderEventListener integration tests

* Removed unused methods in PrisonApiMockServer

Methods used in tests to mock HTTP requests to get
prisoner info and count of requests to Prison API
are no longer used.

Prisoner Search API is used instead.

* Removed Prison API methods to get prisoner info

and the associated types and test helper

* Moved PrisonerAlert class under dto.prisonersearch

Prisoner info are now retrieved using Prisonser Search API
and that includes their alerts.

* Formatted code

* Moved PrisonerNotFoundException with the others

* PrisonerSearchService: Removed trailing '&'

It's unnecessary, we can save the byte :)

* Removed unnecessary `// language=json`

* Continue using Prison API to get prisoner info **by bookingId**

Prisoner Search API is much faster and nicer but only contains
information about the latest bookingId (a prisoner can have
many booking ids, e.g. if they've been in prison multiple
times).

Although this may not be a problem in practive in theory a request
to our API could provide an older booking ID and that would not
work the same way if we start using Prisoner Search API.

For this reason this reverts some of the changes related to getting
information by `bookingId`.

NOTE: The `PrisonerIncentiveReviewService` class handles both
the case where a `prisonerNumber` or a `bookingId` are used.
In the past the information about a prisoner came from Prison API
and the response class was used as a generic input for the methods
to add incentive reviews.

This is no longer the case as data may come from Prisoner Search API
(when getting info by `prisonerNumber`) or from Prison API (when
getting info by `bookingId`). For this reason I've introduced an
`PrisonerBasicInfo` interface.

* Moved PrisonerAlert DTO class directly under dto package

It is now used both in Prison API and Prisoner Search API
responses so it makes sense to move it up a level.

* Moved `PrisonerBasicInfo` interface directly under dto

Again, this is used by Prison API and Prisoner Search API
classes.